Fix starting errors
# Conflicts:
#	src/main/java/be/fgov/ehealth/agreement/protocol/v1/AskAgreementResponse.java
#	src/main/java/be/fgov/ehealth/agreement/protocol/v1/ConsultAgreementRequest.java
#	src/main/java/be/fgov/ehealth/agreement/protocol/v1/ConsultAgreementResponse.java
#	src/main/java/be/fgov/ehealth/agreement/protocol/v1/ObjectFactory.java
#	src/main/kotlin/org/taktik/freehealth/middleware/service/impl/AgreementServiceUtils.kt
