+++ src/main/java/be/fgov/ehealth/agreement/protocol/v1/AskAgreementResponse.java
@@ -1,0 +1,15 @@
+package be.fgov.ehealth.agreement.protocol.v1;
+
+import be.fgov.ehealth.mycarenet.commons.protocol.v4.SendResponseType;
+import javax.xml.bind.annotation.XmlRootElement;
+
+@XmlRootElement(
+   name = "AskAgreementResponse",
+   namespace = "urn:be:fgov:ehealth:mycarenet:agreement:protocol:v1"
+)
+public class AskAgreementResponse extends SendResponseType {
+   private static final long serialVersionUID = 1L;
+
+   public AskAgreementResponse() {
+   }
+}
