+++ src/main/java/be/fgov/ehealth/agreement/protocol/v1/ConsultAgreementRequest.java
@@ -1,0 +1,15 @@
+package be.fgov.ehealth.agreement.protocol.v1;
+
+import be.fgov.ehealth.mycarenet.commons.protocol.v4.SendRequestType;
+import javax.xml.bind.annotation.XmlRootElement;
+
+@XmlRootElement(
+   name = "ConsultAgreementRequest",
+   namespace = "urn:be:fgov:ehealth:mycarenet:agreement:protocol:v1"
+)
+public class ConsultAgreementRequest extends SendRequestType {
+   private static final long serialVersionUID = 1L;
+
+   public ConsultAgreementRequest() {
+   }
+}
