+++ src/main/java/be/fgov/ehealth/agreement/protocol/v1/ConsultAgreementResponse.java
@@ -1,0 +1,15 @@
+package be.fgov.ehealth.agreement.protocol.v1;
+
+import be.fgov.ehealth.mycarenet.commons.protocol.v4.SendResponseType;
+import javax.xml.bind.annotation.XmlRootElement;
+
+@XmlRootElement(
+   name = "ConsultAgreementResponse",
+   namespace = "urn:be:fgov:ehealth:mycarenet:agreement:protocol:v1"
+)
+public class ConsultAgreementResponse extends SendResponseType {
+   private static final long serialVersionUID = 1L;
+
+   public ConsultAgreementResponse() {
+   }
+}
