+++ src/main/kotlin/org/taktik/connector/business/agreement/exception/AgreementBusinessConnectorExceptionValues.kt
@@ -1,21 +1,6 @@
-package org.taktik.connector.business.agreement.exception;
+package org.taktik.connector.business.agreement.exception
 
-public enum AgreementBusinessConnectorExceptionValues {
-   PARAMETER_NULL("parameters.null", "This parameter is null : {0}");
+enum class AgreementBusinessConnectorExceptionValues(val errorCode: String, val message: String) {
+    PARAMETER_NULL("parameters.null", "This parameter is null : {0}")
 
-   private String errorCode;
-   private String message;
-
-   private AgreementBusinessConnectorExceptionValues(String errorCode, String errorMessage) {
-      this.errorCode = errorCode;
-      this.message = errorMessage;
-   }
-
-   public String getErrorCode() {
-      return this.errorCode;
-   }
-
-   public String getMessage() {
-      return this.message;
-   }
 }
