+++ src/main/kotlin/org/taktik/connector/business/agreement/service/ServiceFactory.kt
@@ -1,29 +1,29 @@
-package org.taktik.connector.business.agreement.service;
+package org.taktik.connector.business.agreement.service
 
-import java.util.ArrayList;
-import java.util.List;
-import org.apache.commons.lang.Validate;
-import org.taktik.connector.technical.config.ConfigFactory;
-import org.taktik.connector.technical.config.Configuration;
-import org.taktik.connector.technical.exception.TechnicalConnectorException;
-import org.taktik.connector.technical.service.sts.security.SAMLToken;
-import org.taktik.connector.technical.ws.domain.GenericRequest;
-import org.taktik.connector.technical.ws.domain.TokenType;
+import org.apache.commons.lang.Validate
+import org.taktik.connector.technical.config.ConfigFactory
+import org.taktik.connector.technical.config.Configuration
+import org.taktik.connector.technical.exception.TechnicalConnectorException
+import org.taktik.connector.technical.service.sts.security.SAMLToken
+import org.taktik.connector.technical.ws.domain.GenericRequest
+import org.taktik.connector.technical.ws.domain.TokenType
 
-public final class ServiceFactory {
-   private static final String PROP_ENDPOINT_AGREEMENT = "endpoint.agreement";
-   private static final List<String> expectedProps = new ArrayList();
-   private static final Configuration config;
+object ServiceFactory {
+    private const val PROP_ENDPOINT_AGREEMENT = "endpoint.agreement"
+    private val expectedProps: List<String?> = listOf();
+    private var config: Configuration? = null
+    @Throws(TechnicalConnectorException::class)
+    fun getAgreementPort(token: SAMLToken?): GenericRequest {
+        Validate.notNull(token, "Required parameter SAMLToken is null.")
+        return GenericRequest().setEndpoint(
+            config!!.getProperty(
+                PROP_ENDPOINT_AGREEMENT,
+                "\$uddi{uddi:ehealth-fgov-be:business:mycareneteagreement:v1}"
+            )
+        ).setCredential(token, TokenType.SAML).addDefaulHandlerChain()
+    }
 
-   private ServiceFactory() {
-   }
-
-   public static GenericRequest getAgreementPort(SAMLToken token) throws TechnicalConnectorException {
-      Validate.notNull(token, "Required parameter SAMLToken is null.");
-      return (new GenericRequest()).setEndpoint(config.getProperty(PROP_ENDPOINT_AGREEMENT, "$uddi{uddi:ehealth-fgov-be:business:mycareneteagreement:v1}")).setCredential(token, TokenType.SAML).addDefaulHandlerChain();
-   }
-
-   static {
-      config = ConfigFactory.getConfigValidator(expectedProps);
-   }
+    init {
+        config = ConfigFactory.getConfigValidator(expectedProps)
+    }
 }
