+++ src/main/kotlin/org/taktik/connector/business/agreement/validator/AgreementXmlValidatorImpl.java
@@ -9,8 +9,8 @@ import be.fgov.ehealth.mycarenet.commons.core.v3.CommonInputType;
 import be.fgov.ehealth.mycarenet.commons.core.v3.OriginType;
 import be.fgov.ehealth.mycarenet.commons.core.v3.RequestType;
 import be.fgov.ehealth.mycarenet.commons.core.v3.RoutingType;
-import be.fgov.ehealth.mycarenet.commons.protocol.v3.SendRequestType;
-import be.fgov.ehealth.mycarenet.commons.protocol.v3.SendResponseType;
+import be.fgov.ehealth.mycarenet.commons.protocol.v4.SendRequestType;
+import be.fgov.ehealth.mycarenet.commons.protocol.v4.SendResponseType;
 import org.taktik.connector.technical.validator.impl.XMLValidatorImpl;
 
 public class AgreementXmlValidatorImpl extends XMLValidatorImpl {
