+++ src/main/kotlin/org/taktik/freehealth/middleware/service/impl/AgreementServiceImpl.kt
@@ -2,38 +2,26 @@ package org.taktik.freehealth.middleware.service.impl
 
 import be.cin.encrypted.BusinessContent
 import be.cin.encrypted.EncryptedKnownContent
-import be.fgov.ehealth.aa.complextype.v1.Speciality
 import be.fgov.ehealth.agreement.protocol.v1.AskAgreementRequest
+import be.fgov.ehealth.agreement.protocol.v1.AskAgreementResponse
 import be.fgov.ehealth.chap4.protocol.v1.AskChap4MedicalAdvisorAgreementResponse
 import be.fgov.ehealth.etee.crypto.utils.KeyManager
-import be.fgov.ehealth.messageservices.mycarenet.core.v1.SendTransactionRequest
-import be.fgov.ehealth.mycarenet.commons.core.v4.CareProviderType
-import be.fgov.ehealth.mycarenet.commons.core.v4.CareReceiverIdType
-import be.fgov.ehealth.mycarenet.commons.core.v4.CommonInputType
-import be.fgov.ehealth.mycarenet.commons.core.v4.IdType
-import be.fgov.ehealth.mycarenet.commons.core.v4.LicenseType
-import be.fgov.ehealth.mycarenet.commons.core.v4.NihiiType
-import be.fgov.ehealth.mycarenet.commons.core.v4.OriginType
-import be.fgov.ehealth.mycarenet.commons.core.v4.PackageType
-import be.fgov.ehealth.mycarenet.commons.core.v4.RequestType
-import be.fgov.ehealth.mycarenet.commons.core.v4.RoutingType
-import be.fgov.ehealth.mycarenet.commons.core.v4.ValueRefString
+import be.fgov.ehealth.mycarenet.commons.core.v4.*
 import be.fgov.ehealth.technicalconnector.signature.AdvancedElectronicSignatureEnumeration
 import be.fgov.ehealth.technicalconnector.signature.SignatureBuilderFactory
 import be.fgov.ehealth.technicalconnector.signature.transformers.EncapsulationTransformer
 import org.apache.commons.codec.binary.Base64
 import org.apache.commons.lang.StringUtils
 import org.joda.time.DateTime
+import org.joda.time.format.DateTimeFormat
 import org.slf4j.LoggerFactory
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Service
 import org.taktik.connector.business.agreement.exception.AgreementBusinessConnectorException
-import org.taktik.connector.business.agreement.service.impl.AgreementServiceImpl
-import org.taktik.connector.business.agreement.validator.AgreementXmlValidatorImpl
 import org.taktik.connector.business.domain.agreement.AgreementResponse
-import org.taktik.connector.business.mycarenet.attest.domain.InputReference
-import org.taktik.connector.business.mycarenet.attestv3.mappers.BlobMapper
+import org.taktik.connector.business.mycarenetcommons.mapper.v4.BlobMapper
 import org.taktik.connector.business.mycarenetdomaincommons.builders.BlobBuilderFactory
+import org.taktik.connector.business.mycarenetdomaincommons.domain.InputReference
 import org.taktik.connector.business.mycarenetdomaincommons.util.McnConfigUtil
 import org.taktik.connector.technical.config.ConfigFactory
 import org.taktik.connector.technical.exception.SoaErrorException
@@ -51,9 +39,7 @@ import org.taktik.connector.technical.utils.CertificateParser
 import org.taktik.connector.technical.utils.ConnectorXmlUtils
 import org.taktik.connector.technical.utils.IdentifierType
 import org.taktik.connector.technical.utils.MarshallerHelper
-import org.taktik.connector.technical.validator.XMLValidator
 import org.taktik.freehealth.middleware.dao.User
-import org.taktik.freehealth.middleware.dto.HealthcareParty
 import org.taktik.freehealth.middleware.dto.mycarenet.CommonOutput
 import org.taktik.freehealth.middleware.dto.mycarenet.MycarenetConversation
 import org.taktik.freehealth.middleware.dto.mycarenet.MycarenetError
@@ -61,7 +47,6 @@ import org.taktik.freehealth.middleware.exception.MissingTokenException
 import org.taktik.freehealth.middleware.service.AgreementService
 import org.taktik.freehealth.middleware.service.STSService
 import org.taktik.icure.fhir.entities.r4.Meta
-import org.taktik.icure.fhir.entities.r4.Quantity
 import org.taktik.icure.fhir.entities.r4.attachment.Attachment
 import org.taktik.icure.fhir.entities.r4.binary.Binary
 import org.taktik.icure.fhir.entities.r4.bundle.Bundle
@@ -73,8 +58,6 @@ import org.taktik.icure.fhir.entities.r4.claim.ClaimSupportingInfo
 import org.taktik.icure.fhir.entities.r4.codeableconcept.CodeableConcept
 import org.taktik.icure.fhir.entities.r4.coding.Coding
 import org.taktik.icure.fhir.entities.r4.count.Count
-import org.taktik.icure.fhir.entities.r4.coverage.Coverage
-import org.taktik.icure.fhir.entities.r4.extension.Extension
 import org.taktik.icure.fhir.entities.r4.humanname.HumanName
 import org.taktik.icure.fhir.entities.r4.identifier.Identifier
 import org.taktik.icure.fhir.entities.r4.messageheader.MessageHeader
@@ -89,7 +72,6 @@ import org.taktik.icure.fhir.entities.r4.practitioner.Practitioner
 import org.taktik.icure.fhir.entities.r4.practitionerrole.PractitionerRole
 import org.taktik.icure.fhir.entities.r4.reference.Reference
 import org.taktik.icure.fhir.entities.r4.servicerequest.ServiceRequest
-import org.w3._2005._05.xmlmime.Base64Binary
 import org.w3c.dom.Document
 import org.w3c.dom.NodeList
 import java.io.StringWriter
@@ -105,11 +87,8 @@ import javax.xml.transform.stream.StreamResult
 
 @Service
 class AgreementServiceImpl(private val stsService: STSService, private val keyDepotService: KeyDepotService) : AgreementService {
-    private val agreementServiceUtils = AgreementServiceUtils();
-    private val freehealthAgreementService: org.taktik.connector.business.agreement.service.AgreementService =
-        AgreementServiceImpl()
+    private val freehealthAgreementService: org.taktik.connector.business.agreement.service.AgreementService = org.taktik.connector.business.agreement.service.impl.AgreementServiceImpl()
 
-    private val agreementXmlValidator: XMLValidator = AgreementXmlValidatorImpl()
     private val config = ConfigFactory.getConfigValidator(emptyList())
     private var configValidator = ConfigFactory.getConfigValidator(listOf(
         "chapterIV.keydepot.application",
@@ -136,7 +115,7 @@ class AgreementServiceImpl(private val stsService: STSService, private val keyDe
         return error
     }
 
-    private fun generateError(e: SoaErrorException): org.taktik.connector.business.domain.agreement.AgreementResponse {
+    private fun generateError(e: SoaErrorException): AgreementResponse {
         val error = AgreementResponse()
         error.isAcknowledged = false
         error.errors = Arrays.asList(MycarenetError(code = e.errorCode, msgFr = e.message, msgNl = e.message))
@@ -258,7 +237,7 @@ class AgreementServiceImpl(private val stsService: STSService, private val keyDe
                         }
                     origin = OriginType().apply {
                         `package` = PackageType().apply {
-                            license = LicenseType().apply {
+                            license = be.fgov.ehealth.mycarenet.commons.core.v4.LicenseType().apply {
                                 username = packageInfo.userName
                                 password = packageInfo.password
                             }
@@ -299,26 +278,14 @@ class AgreementServiceImpl(private val stsService: STSService, private val keyDe
             } catch (e: SoaErrorException) {
                 return generateError(e).apply {
                     val rt = e.responseType
-                    if (rt is AskChap4MedicalAdvisorAgreementResponse) {
-                        val co = rt.commonOutput
-                        this.commonOutput = CommonOutput().apply {
-                            this.inputReference = co?.inputReference ?: rt?.recordCommonOutput?.inputReference?.toString()
-                            this.nipReference = co?.nipReference
-                            this.outputReference = co?.outputReference ?: rt?.recordCommonOutput?.outputReference?.toString()
-                        }
-                        this.mycarenetConversation = MycarenetConversation().apply {
-                            rt.soapRequest?.writeTo(this.soapRequestOutputStream())
-                            rt.soapRequest?.writeTo(this.soapResponseOutputStream())
-                        }
-                        rt.returnInfo?.let { ri ->
-                            this.errors = this.errors?.let { it + listOf(MycarenetError(code = ri.faultCode, path = ri.faultSource, msgFr = ri.message.value, msgNl = ri.message.value))}
-                        }
+                    if (rt is AskAgreementResponse) {
+                        // do something
                     }
                 }
             }
 
             val commonOutput =
-                CommonOutput(response.`return`.commonOutput?.inputReference, response.`return`.commonOutput.nipReference, response.`return`.commonOutput.outputReference)
+                CommonOutput(response?.`return`?.commonOutput?.inputReference, response?.`return`?.commonOutput?.nipReference, response?.`return`?.commonOutput?.outputReference)
 
             return AgreementResponse().apply {
 
@@ -354,7 +321,7 @@ class AgreementServiceImpl(private val stsService: STSService, private val keyDe
         numberOfSessionForAnnex2: Float?
     ): AskAgreementRequest{
 
-        val claim = agreementServiceUtils.getClaim(
+        val claim = this.getClaim(
             claimId = "1",
             claimStatus = "active",
             subTypeCode = "kine",
@@ -365,7 +332,7 @@ class AgreementServiceImpl(private val stsService: STSService, private val keyDe
             providerType = ""
         )
 
-        val bundle = agreementServiceUtils.getBundle(requestType, claim, messageEventSystem, messageEventCode, patientFirstName, patientLastName, patientGender, patientSsin, patientIo, patientIoMembership, hcpNihii, hcpFirstName, hcpLastName, orgNihii, organizationType, annex1, annex2, parameterNames, agreementStartDate, agreementEndDate, agreementType, numberOfSessionForAnnex1, numberOfSessionForAnnex2)
+        val bundle = this.getBundle(requestType, claim, messageEventSystem, messageEventCode, patientFirstName, patientLastName, patientGender, patientSsin, patientIo, patientIoMembership, hcpNihii, hcpFirstName, hcpLastName, orgNihii, organizationType, annex1, annex2, parameterNames, agreementStartDate, agreementEndDate, agreementType, numberOfSessionForAnnex1, numberOfSessionForAnnex2)
         return AskAgreementRequest().apply {
 
 
@@ -404,26 +371,14 @@ class AgreementServiceImpl(private val stsService: STSService, private val keyDe
         } catch (e: SoaErrorException) {
             return generateError(e).apply {
                 val rt = e.responseType
-                if (rt is AskChap4MedicalAdvisorAgreementResponse) {
-                    val co = rt.commonOutput
-                    this.commonOutput = CommonOutput().apply {
-                        this.inputReference = co?.inputReference ?: rt?.recordCommonOutput?.inputReference?.toString()
-                        this.nipReference = co?.nipReference
-                        this.outputReference = co?.outputReference ?: rt?.recordCommonOutput?.outputReference?.toString()
-                    }
-                    this.mycarenetConversation = MycarenetConversation().apply {
-                        rt.soapRequest?.writeTo(this.soapRequestOutputStream())
-                        rt.soapRequest?.writeTo(this.soapResponseOutputStream())
-                    }
-                    rt.returnInfo?.let { ri ->
-                        this.errors = this.errors?.let { it + listOf(MycarenetError(code = ri.faultCode, path = ri.faultSource, msgFr = ri.message.value, msgNl = ri.message.value))}
-                    }
+                if (rt is AskAgreementResponse) {
+                    // do something
                 }
             }
         }
 
         val commonOutput =
-            CommonOutput(response.`return`.commonOutput?.inputReference, response.`return`.commonOutput.nipReference, response.`return`.commonOutput.outputReference)
+            CommonOutput(response?.`return`?.commonOutput?.inputReference, response?.`return`?.commonOutput?.nipReference, response?.`return`?.commonOutput?.outputReference)
 
         return AgreementResponse().apply {
 
@@ -502,6 +457,530 @@ class AgreementServiceImpl(private val stsService: STSService, private val keyDe
         }
 
         return sb.toString()
+    }
+
+    fun getPractitionerRole(practitionerRoleId: String, hcpNihii: String, hcpFirstName: String, hcpLastName: String): PractitionerRole {
+        return PractitionerRole().apply {
+            id = "PractitionerRole$practitionerRoleId"
+            meta = Meta(
+                profile = listOf("https://www.ehealth.fgov.be/standards/fhir/core/StructureDefinition/be-practitionerrole")
+            )
+            practitioner = Reference().apply {
+                reference = "Practitioner/Practitioner$practitionerRoleId"
+            }
+            code = listOf(
+                CodeableConcept().apply {
+                    coding = listOf(
+                        Coding(
+                            system = "https://www.ehealth.fgov.be/standards/fhir/core/CodeSystem/cd-hcparty",
+                            code = "persphysiotherapist"
+                        ))
+                }
+            )
+        }
+    }
+
+    fun getPractitioner(practitionerId: String, hcpNihii: String, hcpFirstName: String, hcpLastName: String): Practitioner {
+        return Practitioner().apply {
+            id = "Practitioner$practitionerId"
+            meta = Meta(
+                profile = listOf("https://www.ehealth.fgov.be/standards/fhir/core/StructureDefinition/be-practitioner")
+            )
+            identifier = listOf(
+                Identifier().apply {
+                    system = "https://www.ehealth.fgov.be/standards/fhir/core/NamingSystem/nihdi"
+                    value = hcpNihii
+                }
+            )
+            name = listOf(
+                HumanName(
+                    family = hcpLastName,
+                    given = listOf(hcpFirstName)
+                )
+            )
+        }
+    }
+
+    fun getCodableConcept(system: String, code: String): CodeableConcept{
+        return CodeableConcept().apply {
+            coding = listOf(
+                Coding(
+                    system = system,
+                    code = code
+                )
+            )
+        }
+    }
+
+    fun getClaim(
+        claimId: String,
+        claimStatus: String,
+        subTypeCode: String,
+        agreementStartDate: DateTime,
+        insuranceRef: String,
+        pathologyCode: String,
+        pathologyStartDate: DateTime?,
+        providerType: String
+    ): Claim {
+        return Claim(
+            patient = Reference().apply { reference = "Patient/Patient1" },
+            priority = getCodableConcept("http://terminology.hl7.org/CodeSystem/processpriority", "stat"),
+            provider = Reference().apply { reference = providerType },
+            type = getCodableConcept("http://terminology.hl7.org/CodeSystem/claim-type", "professional")
+        ).apply {
+            id = "Claim$claimId"
+            meta = Meta(
+                profile = listOf("https://www.ehealth.fgov.be/standards/fhir/mycarenet/StructureDefinition/be-eagreementclaim-kine")
+            )
+            status = claimStatus
+            subType = getCodableConcept("https://www.ehealth.fgov.be/standards/fhir/mycarenet/CodeSystem/agreement-types", subTypeCode)
+            use = "preauthorization"
+            billablePeriod = getBillablePeriod(agreementStartDate!!)
+            created = DateTime().toString()
+            enterer = Reference().apply { reference = "PractitionerRole/PractitionerRole1"}
+            referral = Reference().apply {
+                reference = "ServiceRequest/ServiceRequest1"
+            }
+            insurance = listOf(getInsurance(insuranceRef, ""))
+            supportingInfo = listOf(
+                getSupportingInfo(1, "attachment", "functional-report", null, null, "QW5uZXhlIGlubGluZSwgYmFzZTY0ZWQ=", "nom/description de l'annexe", "application/pdf"),
+                getSupportingInfo(2, "info", null, null, "additional Information", null, null, null),
+                getSupportingInfo(3, "info", null, "ServiceRequest/ServiceRequest2", null, null, null, null)
+            )
+
+            item = listOf(getServicedDateItem(pathologyStartDate!!, pathologyCode, 1), getCodeItem(pathologyCode))
+        }
+    }
+
+    fun getSupportingInfo(sequenceNumber: Int, claimInformationCategory: String, annexTypeCode: String?, valueReference: String?, valueString: String?, valueAttachmentData: String?, valueAttachmentTitle: String?, valueAttachmentContentType: String?): ClaimSupportingInfo {
+        var code: CodeableConcept? = null
+        val category = CodeableConcept().apply {
+            coding = listOf(Coding(
+                system = "http://terminology.hl7.org/CodeSystem/claiminformationcategory",
+                code = claimInformationCategory
+            ))
+        }
+
+        if(!annexTypeCode.isNullOrEmpty()){
+            code = CodeableConcept().apply {
+                coding = listOf(Coding(
+                    system = "https://www.ehealth.fgov.be/standards/fhir/mycarenet/CodeSystem/annex-types",
+                    code = annexTypeCode
+                ))
+            }
+        }
+
+
+        return ClaimSupportingInfo(
+            category = category
+        ).apply {
+            when{
+                code != null -> this.code = code
+                !valueReference.isNullOrEmpty() -> this.valueReference = Reference().apply {
+                    reference = valueReference
+                }
+                !valueString.isNullOrEmpty() -> this.valueString = valueString
+                !valueAttachmentData.isNullOrEmpty() -> this.valueAttachment = Attachment().apply {
+                    contentType = valueAttachmentContentType
+                    data = valueAttachmentData
+                    title = valueAttachmentTitle
+                }
+            }
+
+            sequence = sequenceNumber
+
+        }
+    }
+
+    fun getOrganization(organizationId: String, orgNihii: String, organizationType: String): Organization {
+        return Organization().apply {
+            id = "Organization$organizationId"
+            meta = Meta(
+                profile = listOf("https://www.ehealth.fgov.be/standards/fhir/core/StructureDefinition/be-organization")
+            )
+            identifier = listOf(
+                Identifier().apply {
+                    system = "https://www.ehealth.fgov.be/standards/fhir/core/NamingSystem/nihdi"
+                    value = orgNihii
+                }
+            )
+            type = listOf(
+                CodeableConcept().apply {
+                    coding = listOf(
+                        Coding(
+                            system = "https://www.ehealth.fgov.be/standards/fhir/core/CodeSystem/cd-hcparty",
+                            code = organizationType
+                        )
+                    )
+                }
+            )
+        }
+    }
+
+    fun getPatient(patientFirstName: String, patientLastName: String, gender: String, patientSsin: String?, io: String ?, ioMembership: String?): Patient {
+        return Patient(
+            id = "Patient1",
+            meta = Meta(
+                profile = listOf("https://www.ehealth.fgov.be/standards/fhir/core/StructureDefinition/be-patient")
+            ),
+            identifier = listOf(
+                Identifier().apply {
+                    when{
+                        !patientSsin.isNullOrEmpty() && ioMembership.isNullOrEmpty() -> {
+                            system = "https://www.ehealth.fgov.be/standards/fhir/core/NamingSystem/ssin"
+                            value = patientSsin
+                        }
+                        patientSsin.isNullOrEmpty() && !ioMembership.isNullOrEmpty() && !io.isNullOrEmpty() -> {
+                            system = "https://www.ehealth.fgov.be/standards/fhir/core/NamingSystem/insurancymembership"
+                            value = ioMembership
+                            assigner = Reference(identifier = Identifier().apply {
+                                system = "https://www.ehealth.fgov.be/standards/fhir/core/NamingSystem/insurancenumber"
+                                value = io
+                            })
+                        }
+                        !patientSsin.isNullOrEmpty() && !ioMembership.isNullOrEmpty() -> {
+                            system = "https://www.ehealth.fgov.be/standards/fhir/core/NamingSystem/ssin"
+                            value = patientSsin
+                        }
+                    }
+                }),
+            name = listOf(
+                HumanName(
+                    family = patientLastName,
+                    given = listOf(patientFirstName)
+                )
+            ),
+            gender = gender
+        )
+    }
+
+    fun getServiceRequest(serviceRequestId: String, data: String, annexId: String, quantity: Float, patientFirstName: String, patientLastName: String, gender: String, hcpNihii: String, hcpFirstName: String, hcpLastName: String, patientSsin: String?, io: String?, ioMembership: String?): ServiceRequest {
+        val patient = getPatient(patientFirstName!!, patientLastName!!, gender!!, patientSsin, io, ioMembership)
+        val formatter = DateTimeFormat.forPattern("yyyy-MM-dd")
+        return ServiceRequest(subject = Reference().apply { patient }).apply {
+            id = "ServiceRequest$serviceRequestId"
+            meta = Meta(
+                profile = listOf("https://www.ehealth.fgov.be/standards/fhir/mycarenet/StructureDefinition/be-eagreementservicerequest")
+            )
+            contained = getContained(data, annexId)
+            identifier = listOf(
+                Identifier().apply {
+                    system = "https://www.ehealth.fgov.be/standards/fhir/core/NamingSystem/uhmep"
+                    value = "Prescription id"
+                }
+            )
+            status = "active"
+            intent = "order"
+            category = listOf(
+                CodeableConcept().apply {
+                    coding = listOf(
+                        Coding(
+                            system = "http://snomed.info/sct",
+                            code = "91251008",
+                            display = "Physical therapy procedure"
+                        )
+                    )
+                }
+            )
+            code = CodeableConcept().apply {
+                coding = listOf(
+                    Coding(
+                        system = "http://snomed.info/sct",
+                        code = "91251008",
+                        display = "Physical therapy procedure"
+                    )
+                )
+            }
+            quantityQuantity = Count().apply {
+                value = quantity
+            }
+            subject = Reference().apply {
+                reference = "Patient/Patient1"
+            }
+            authoredOn = formatter.print(DateTime())
+            requester = Reference().apply {
+                reference = "PractitionerRole/PractitionerRole2"
+            }
+            supportingInfo = listOf(Reference().apply {
+                reference = "#annexSR$annexId"
+            })
+        }
+    }
+
+    fun getContained(data: String, containedId: String): List<Binary>{
+        return listOf(
+            Binary(
+                contentType = "application/pdf",
+                data = data,
+                id = "annexSR$containedId"
+            )
+        )
+    }
+
+    fun getParameters(parameterId: String, parameterNames: Array<String>,
+                      agreementTypes: String,
+                      startDate: DateTime?,
+                      endDate: DateTime?,
+                      patientFirstName: String?,
+                      patientLastName: String?,
+                      patientGender: String?,
+                      patientSsin: String?,
+                      io: String?,
+                      ioMembership: String?): Parameters {
+        val param = mutableListOf<ParametersParameter>();
+        for (parameterName in parameterNames){
+            param.add(getParameter(parameterName, agreementTypes, startDate, endDate, patientFirstName, patientLastName, patientGender, patientSsin, io, ioMembership))
+        }
+        return Parameters().apply {
+            id = "Parameters$parameterId"
+            parameter = param
+        }
+    }
+
+    fun getParameter(parameterName: String,
+                     agreementTypes: String?,
+                     startDate: DateTime?,
+                     endDate: DateTime?,
+                     patientFirstName: String?,
+                     patientLastName: String?,
+                     patientGender: String?,
+                     patientSsin: String?,
+                     io: String?,
+                     ioMembership: String?
+    ): ParametersParameter{
+        return ParametersParameter().apply {
+            name = parameterName
+            when{
+                parameterName == "resourceType" -> valueString = "ClaimResponse"
+                parameterName == "patient" -> valueReference = Reference().apply {
+                    reference = "Patient/Patient1"
+                }
+                parameterName == "use" -> valueCode = "preauthorization"
+                parameterName == "subType" -> valueCoding = Coding().apply {
+                    system = "https://www.ehealth.fgov.be/standards/fhir/mycarenet/CodeSystem/agreement-types"
+                    code = agreementTypes
+                }
+                parameterName == "preAuthPeriod" -> valuePeriod = Period().apply {
+                    start = startDate.toString()
+                    end = endDate.toString()
+                }
+            }
+        }
+    }
+
+    fun getInsurance(insuranceRef: String, display: String): ClaimInsurance{
+        return ClaimInsurance(
+            sequence = 1,
+            focal = true,
+            coverage = Reference(
+                display = display
+            ),
+            preAuthRef = listOf(insuranceRef)
+        )
+    }
+
+    fun getBillablePeriod(startDate: DateTime): Period{
+        val formatter = DateTimeFormat.forPattern("yyyy-MM-dd")
+        return Period(
+            start = formatter.print(startDate)
+        )
+    }
+
+    fun getServicedDateItem(pathologyDate: DateTime, pathologyCode: String, sequenceNumber: Int): ClaimItem {
+        val formatter = DateTimeFormat.forPattern("yyyy-MM-dd")
+        return ClaimItem(
+            sequence = sequenceNumber,
+            productOrService = CodeableConcept(
+                coding = listOf(
+                    Coding(
+                        system = "https://www.ehealth.fgov.be/standards/fhir/mycarenet/CodeSystem/nihdi-physiotherapy-pathologysituationcode",
+                        code = pathologyCode
+                    )
+                )
+            ),
+            servicedDate = formatter.print(pathologyDate)
+        )
+    }
+
+    fun getCodeItem(code: String): ClaimItem {
+        return ClaimItem(
+            productOrService = getCodableConcept("https://www.ehealth.fgov.be/standards/fhir/mycarenet/CodeSystem/nihdi-physiotherapy-pathologysituationcode", code)
+        )
+    }
+
+    fun getAdditionalNotes(additionalNotes: String): ClaimSupportingInfo{
+        return ClaimSupportingInfo(
+            sequence = 2,
+            category = CodeableConcept().apply {
+                coding = listOf(Coding().apply {
+                    system = "http://terminology.hl7.org/CodeSystem/claiminformationcategory"
+                    code = "info"
+                })
+            },
+            valueString = additionalNotes
+        )
+    }
+
+    fun getPrescriptionInfos(quantity: Int, document: String, documentId: String): ServiceRequest{
+        return ServiceRequest(
+            identifier = listOf(
+                Identifier(
+                    value = "Prescription id",
+                    system = "https://www.ehealth.fgov.be/standards/fhir/core/NamingSystem/uhmep"
+                )
+            ),
+            status = "active",
+            intent = "order",
+            category = listOf(
+                CodeableConcept(
+                    coding = listOf(
+                        Coding(
+                            system = "http://snomed.info/sct",
+                            code = "procedure snomed",
+                            display = "?????"
+                        )
+                    )
+                )
+            ),
+            code = CodeableConcept(
+                coding = listOf(
+                    Coding(
+                        system = "http://snomed.info/sct",
+                        code = "procedure snomed",
+                        display = "?????"
+                    )
+                )
+            ),
+            quantityQuantity = Count().apply {
+                value = quantity.toFloat()
+            },
+            subject = Reference().apply {
+                reference = "Patient/Patient1"
+            },
+            authoredOn = DateTime.now().toString("yyyy-MM-dd"),
+            requester = Reference().apply {
+                PractitionerRole(
+                    practitioner = Reference().apply {
+                        reference = "PractitionerRole/PractitionerRole2"
+                    }
+                )
+            },
+            supportingInfo = listOf(Reference().apply {
+                Binary(
+                    contentType = "application/pdf",
+                    data = document,
+                    id = documentId
+                )
+            })
+        )
+    }
+
+    fun getMessageHeader(claim: Claim, messageEventSystem: String, messageEventsCode: String): MessageHeader {
+        return MessageHeader(
+            eventCoding = Coding(
+                system = messageEventSystem,
+                code = messageEventsCode
+            ),
+            source = MessageHeaderSource()
+        ).apply {
+            id = "generate id"
+            meta = Meta(
+                profile = listOf("https://www.ehealth.fgov.be/standards/fhir/mycarenet/StructureDefinition/be-eagreementmessageheader")
+            )
+
+            destination = listOf(
+                MessageHeaderDestination(
+                    endpoint = "MyCareNet",
+                    name = "MyCareNet"
+                )
+            )
+            focus = listOf(
+                Reference().apply {
+                    claim
+                }
+            )
+        }
+    }
+
+    fun getBundle(
+        requestType: org.taktik.freehealth.middleware.service.impl.AgreementServiceImpl.RequestTypeEnum,
+        claim: Claim,
+        messageEventSystem: String,
+        messageEventCode: String,
+        patientFirstName: String,
+        patientLastName: String,
+        patientGender: String,
+        patientSsin: String?,
+        patientIo: String?,
+        patientIoMembership: String?,
+        hcpNihii: String,
+        hcpFirstName: String,
+        hcpLastName: String,
+        orgNihii: String?,
+        organizationType: String?,
+        annex1: String?,
+        annex2: String?,
+        parameterNames: Array<String>?,
+        agreementStartDate: DateTime?,
+        agreementEndDate: DateTime?,
+        agreementType: String?,
+        numberOfSessionForAnnex1: Float?,
+        numberOfSessionForAnnex2: Float?
+    ): Bundle {
+        return Bundle().apply {
+            id = "Bundle1"
+            meta = Meta(
+                profile = listOf("https://www.ehealth.fgov.be/standards/fhir/mycarenet/StructureDefinition/be-eagreementdemand")
+            )
+            type = "message"
+            timestamp = DateTime().toString()
+            entry = listOf<BundleEntry>().apply {
+                BundleEntry(
+                    resource = getMessageHeader(claim, messageEventSystem, messageEventCode),
+                    fullUrl = "https://www.hl7.org/fhir/bundle-definitions.html#Bundle.entry.fullUrl"
+                )
+                BundleEntry(
+                    resource = getPatient(patientFirstName, patientLastName, patientGender, patientSsin, patientIo, patientIoMembership),
+                    fullUrl = "https://www.hl7.org/fhir/bundle-definitions.html#Bundle.entry.fullUrl"
+                )
+                BundleEntry(
+                    resource = getPractitionerRole("1", hcpNihii, hcpFirstName, hcpLastName),
+                    fullUrl = "https://www.hl7.org/fhir/bundle-definitions.html#Bundle.entry.fullUrl"
+                )
+                BundleEntry(
+                    resource = getPractitioner("1", hcpNihii, hcpFirstName, hcpLastName),
+                    fullUrl = "https://www.hl7.org/fhir/bundle-definitions.html#Bundle.entry.fullUrl"
+                )
+                BundleEntry(
+                    resource = getOrganization("1", orgNihii!!, organizationType!!),
+                    fullUrl = "https://www.hl7.org/fhir/bundle-definitions.html#Bundle.entry.fullUrl"
+                )
+                when{
+                    //Claim 1
+                    requestType != org.taktik.freehealth.middleware.service.impl.AgreementServiceImpl.RequestTypeEnum.CONSULT_LIST -> BundleEntry(
+                        resource = claim,
+                        fullUrl = "https://www.hl7.org/fhir/bundle-definitions.html#Bundle.entry.fullUrl"
+                    )
+                    //Parameters 1
+                    requestType == org.taktik.freehealth.middleware.service.impl.AgreementServiceImpl.RequestTypeEnum.CONSULT_LIST ->   BundleEntry(
+                        resource = getParameters("1", parameterNames!!, agreementType!!, agreementStartDate, agreementEndDate, hcpNihii, hcpFirstName, hcpLastName, patientSsin, patientIo, patientIoMembership),
+                        fullUrl = "https://www.hl7.org/fhir/bundle-definitions.html#Bundle.entry.fullUrl"
+                    )
+                    //Service Request 1
+                    requestType != org.taktik.freehealth.middleware.service.impl.AgreementServiceImpl.RequestTypeEnum.CANCEL && requestType != org.taktik.freehealth.middleware.service.impl.AgreementServiceImpl.RequestTypeEnum.CONSULT_LIST ->
+                        BundleEntry(
+                            resource = getServiceRequest("1", annex1!!, "1", numberOfSessionForAnnex1!!, patientFirstName, patientLastName, patientGender, hcpNihii, hcpFirstName, hcpLastName, patientSsin, patientIo, patientIoMembership),
+                            fullUrl = "https://www.hl7.org/fhir/bundle-definitions.html#Bundle.entry.fullUrl"
+                        )
+                    //Service request 2
+                    requestType == org.taktik.freehealth.middleware.service.impl.AgreementServiceImpl.RequestTypeEnum.ARGUE || requestType == org.taktik.freehealth.middleware.service.impl.AgreementServiceImpl.RequestTypeEnum.ASK ->
+                        BundleEntry(
+                            resource = getServiceRequest("2", annex2!!, "2", numberOfSessionForAnnex2!!, patientFirstName, patientLastName, patientGender, hcpNihii, hcpFirstName, hcpLastName, patientSsin, patientIo, patientIoMembership),
+                            fullUrl = "https://www.hl7.org/fhir/bundle-definitions.html#Bundle.entry.fullUrl"
+                        )
+                }
+            }
+        }
     }
 }
